pyproject.toml
[project]
name = "kq-rag"
version = "0.1.0"
description = "RAG demo for 矿权"
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "lab-1806-vec-db==0.2.2",
    "openai==1.51.0",
    "pip>=24.2",
    "python-dotenv==1.0.1",
    "sentence-transformers==3.1.1",
    "tqdm==4.66.5",
]
